src/ai-agent-service/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1

# AI编码服务依赖
# PyPI当前可用最新稳定版为0.0.20（容器内构建失败即由版本约束导致）
claude-code-sdk==0.0.20
pydantic-settings>=2.0.0

# 额外的日志和调试工具
colorlog>=6.0.0
psutil>=5.9.0  # 用于监控资源使用情况
